added union

+++ user/benchmark.c
@@ -3,17 +3,17 @@
 #include "user/user.h"
 
 void main (int argc, char* argv[]) {
-	int start = uptime_seconds(); 
+	double start = uptime_seconds(); 
 
 	if (fork() == 0)
 		exec(argv[1], ++argv);
 
 	wait(0);
-	int end = uptime_seconds();
+	double end = uptime_seconds();
 
-	printf("It took %d seconds to run\n", end-start); 
+	printf("It took %lf seconds to run\n", end-start); 
 
-	printf("Start: %d \nEnd: %d \n", start, end);
+	printf("Start: %d \nEnd: %lf \n", start, end);
 
 	//  HaHa, theres a security issue here since we don't check for 0 at the end
 	//  TODO: fix the security issue
